requirements.txt
Flask==2.2.5
gunicorn==20.1.0
python-dotenv==1.0.0
bcrypt==4.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
pytest==7.4.0
requests==2.31.0
SQLAlchemy==1.4.49
Flask-SQLAlchemy==3.0.5
Faker==25.2.0
